Add logic to control who can create meetings
Previously any moderator with a moderator role in a room could create
meetings in it. Now there is a method called 'bigbluebutton_can_create?'
that is called to defined whether the user is allowed to create a meeting
in the target room.
Can also be used to execute anything right before a meeting is created
(e.g. adapt metadata values).