Modify empty json responses to prevent errors with jquery
When using 'head :ok' to respond with a json, the body is sent as ' ' (an empty
space). This is interpreted by jquery as an error even if the response code is 200.
The way we're doing it now sends a string in the body and is interpreted as a
successful response. See more about it at:
* http://stackoverflow.com/questions/12407328/rails-head-ok-interpreted-as-ajaxerror
* https://github.com/rails/rails/issues/1742