Update meetings on create, join, end and periodically, refs #1032
Using resque to schedule a job that will update the meetings. The job is
schedule always when one of these calls is made upon a room: create, join
and end.
There's also a rake task that is triggered periodically by cron to check for
meetings that were running and are not anymore.