Created a way to let users change the layout of the rooms.
To do this, a new model was created to receive the options from the room. In this model a user can set the default layout who he wanted. A new function was created in this model to return all the layouts allowed in the server, for now this function just return a list of fixed layouts.

Refs #962