Improve how additional options are sent to create calls
So far, to send parameters to create calls all parameters had to be in the
database. Now there's a new method 'bigbluebutton_create_options' that can
return a hash of parameters that will override the parameters in the database
when sending a create call. Can be used to force some options when creating
a meeting without needing to save it to the database.