New log to join meetings from mobile devices
Now there's no specific page to join from a mobile. Once /join is called,
if the user is in a mobile device (detected using the user's "user-agent")
then a page will be rendered and the user will be automatically redirected
to the conference using the mobile client. This page has also more information
just in case the user is not properly redirected (if the mobile client is
not installed, for example).

Also, there's no more login via QR Code. It wasn't a good solution and was not
being used, so it was simply removed.