Unifying actions 'join' and 'auth' and breaking 'join' in blocks
Now there's a single action to join meetings: 'join'. It can be called via
GET or POST, and accepts the parameters as previously accepted by 'auth'
(user's name and password).
Also, 'join' is now broken in several smaller methods (before filters) to make
it simpler.

This was all done to make the join process in mobile devices better. We need
to display the 'join_mobile' view as an intermediary step when joining from
mobile devices. And it needs to be done when joining directly from 'join' or
from 'invite', and previously was done only for 'join'.

Not finished nor tested yet.

refs #1247