Simplify playback types views/routes, keep only the absolute necessary
We don't need all actions and views for playback types, since this is just
a helper model to store data for recordings. The only thing we really need
is a way to set playback types as shown/hidden.
So kept the update action only and simplified the views to be more like the
way they are intended to be in the apps: a single view where an admin can
update the visibility of all playback types.

refs #1142