Playback formats don't require a playback type, and other small fixes
Allow playback formats to have a nil playback type. It prevents errors if
the data fetched from the web conference server is missing data or has
invalid tags, for example.

Also made BigbluebuttonPlaybackType#name more like the default I18n.t. Now
it returns a "stringified" version of the key searched for if the key does
not exist in the locales.

refs #1142