Don't generate voice bridges, let the webconf server generate it
Now the gem will not set the voice bridge parameter when creating a meeting.
Doing so, the web conference server will generate it for us.
It's the easiest way to solve problems with duplicated voice bridges, an issue
that would happen on an instance of Mconf-Web with 9999 rooms or more.
Since the web conference server knows only of meetings that are running, now
the limit was expanded to 99999 *simultaneous* rooms/meetings.

There is still a parameter in the gem that can be set in case an application
wants to use the voice bridges set in the gem.