Improve lookup for layout names in the locales
Now won't require the locale strings to be placed where defined by the name
of the layout taken from the webconference server. Will store all layouts
in the same place in the locales and lookup for them using only the name
that defines the layout, not the entire name (e.g. only 'defaultlayout'
instead of 'bbb.layout.name.defaultlayout').

refs #1571