Better logic and more tests for auto setting API version
Added several tests that describe how the feature should work. In short,
the version of the API of a server is automatically fetched from the server
if the BigbluebuttonServer doesn't have a version set or if it's empty.
So, to get the version from the server, is enough to just save the server
with version=nil. And the user can still force set a version on it if
he wishes to.

Only does it when updating the server. If a server has no version set for
some reason (e.g. legacy db), it will cause it to fetch the API version
from the server after each request. To end it, just save the server.

refs #1787