First pass improving how BigbluebuttonMeeting's are created
Trying to make it more reliable and with more information. There are
some issues to deal with in the way the gem finds out about meetings
being created and ended, so it's not an entirely simple process.
Now the meeting objects should be created as soon as a meeting is actually
created in the web conference server and there are background jobs
checking for whether they are still running and whether they are still
in memory in the web conference server or not.