Centralize the process of ending meetings after get_meeting_info
Instead of doing it in a worker and in lib/background_tasks, now it is
centralized in the method BigbluebuttonRoom#fetch_meeting_info.
It will check for exceptions and mark the meeting as ended.

Now it also has a way to reopen meetings. It is useful in case the server
stops responding and the meeting is marked as ended. Previously it would
be ended forever and a new meeting record would be created if the server
started responding again.

One last change is that it now considers any BigBlueButtonException to end
the meeting, not just the ones with the key 'notFound'. This works better
now since it can also reopen ended meetings if needed.

Improves PR #126