Don't set recordings as unavailable unless it's a full sync
Was setting recordings as unavailable even if the call to get_recordings
had filters, which meant that the response did not include all recordings
available in the server.
With that, all calls to fetch_recordings that had filters would make all
other recordings in that server (the ones that didn't match the filter)
unavailable.

Closes #132