New worker to fetch recordings after meetings are ended
It will improve the speed with which recordings are found after meetings
end.
Also fixed setting the #recorded attributes in newly created meetings, it
was always using the attributes from model, without considering that it
is possible to override these attributes when making a "create" call (which
is exactly what Mconf-Web does).