No need to call end before a create anymore
A call to /end was originally added by #25 to prevent API errors when
trying to create the same room with different parameters. BigBlueButton
now behaves differently and, even if we change the passwords of a room
when calling /create, the response will be a success and not a failure.

Calling /end was causing issues when multiple users were joining a meeting
too close to each other. The gem wouldn't detect that the meeting was already
started and the second user would kick the first one out.