Fetch recordings for each room instead of each server on sync
When synchronizing recordings, instead of iterating over the servers,
iterate over all rooms. There are usually way more rooms in the DB, but
their getRecordings call are way smaller than getting all recordings in
a server. We're replacing a few (sometimes a single) getRecordings that
might be too big and slow, with several smaller ones.

The old method still exists and can be easily used by applications if
needed.